AI- Talent - Intellegence/requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
groq
pypdf
scikit-learn